pyproject.toml
[project]
name = "bertalign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "googletrans==4.0.0rc1",
    "numba==0.60.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-docx>=1.1.2",
    "sentence-splitter==1.4",
    "sentence-transformers==3.2.1",
]
